from NO3M:
Changes to AutoCQ:

AutoCQ is paused when:

Call field is populated
<Esc> during CQ or countdown
Working an Alt-D QSO

AutoCQ is resumed when:

<Enter>, <F1>, or <F2> is pressed to send next CQ
QSO is logged with normal ESM or CNTRL-Enter (including Alt-D QSO)

AutoCQ is NOT resumed when:

QSO is logged with Shift-Enter (ie. silent logging) or Backslash (quick QSL)

in the above cases, AutoCQ is resumed when <Enter>, <F1>, or <F2> is
pressed to send CQ again.

Bug fixes:

F1/F2 were being captured by AutoCQ instead of sending exchange
message if focus is in Alt-D radio fields.  Function keys were not
changing transmit focus to Alt-D radio if pressed when Alt-D radio
fields have focus and AutoCQ is active.  AutoCQ was being disabled if
using Alt-R or CNTRL-R to switch radios when an Alt-D QSO was in
progress; fixed to just sleep.