patch from NO3M:
Attached patch should fix the AutoCQ being disabled by QSY.  The only
way I could figure to resolve it was to make a autoCQ tracking
variable set to the activeRadio when autoCQ is enabled.  Once the QSY
code is hit in So2sdr::checkSpot, AltDActive would already be reset to
'0', and checking AltDActiveRadio would be a hit and miss if AltD was
never active in the first place.

This bug only happened if QSYing the S/P radio before the autoCQ timer
expired and triggered a CQ on the run radio.