Slight change in enter key processing:
Before, in S&P mode, enter would not log the qso if the call
was never sent. This created problems for me in the Sprint: I would
get a qso totally messed up, then esc out and enter everything
again in S&P mode. After sending the exchange with F2, the qso
would not log because F1 was not sent the second time I entered
the qso.

Does this change mess anything else up?